pyproject.toml
[tool.poetry]
name = "moodify-backend"
version = "0.1.0"
description = "Backend for Moodify Emotion-Based Music Recommendation App"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
absl-py = "2.1.0"
accelerate = "0.34.2"
aiohappyeyeballs = "2.4.0"
aiohttp = "3.10.5"
aiosignal = "1.3.1"
anyio = "4.6.0"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
appnope = "0.1.4"
asgiref = "3.8.1"
asttokens = "2.4.1"
astunparse = "1.6.3"
async-lru = "2.0.4"
async-timeout = "4.0.3"
attrs = "24.2.0"
audioread = "3.0.1"
babel = "2.16.0"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
blinker = "1.8.2"
certifi = "2024.8.30"
cffi = "1.17.1"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
comm = "0.2.2"
contourpy = "1.3.0"
cycler = "0.12.1"
dataclasses = "0.6"
datasets = "3.0.0"
debugpy = "1.8.5"
decorator = "4.4.2"
defusedxml = "0.7.1"
dill = "0.3.8"
dj-rest-auth = "6.0.0"
Django = "5.1.1"
django-allauth = "65.0.1"
django-cors-headers = "4.4.0"
django-redis = "5.4.0"
django-rest-framework-mongoengine = "3.4.1"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
dnspython = "2.6.1"
drf-yasg = "1.21.7"
exceptiongroup = "1.2.2"
executing = "2.1.0"
facenet-pytorch = "2.6.0"
fastjsonschema = "2.20.0"
fer = "22.5.1"
ffmpeg = "1.4"
filelock = "3.16.1"
Flask = "3.0.3"
Flask-Cors = "5.0.0"
flatbuffers = "24.3.25"
fonttools = "4.54.1"
fqdn = "1.5.1"
frozenlist = "1.4.1"
fsspec = "2024.6.1"
gast = "0.6.0"
google-pasta = "0.2.0"
gdown = "5.2.0"
grpcio = "1.66.1"
gunicorn = "23.0.0"
h11 = "0.14.0"
h5py = "3.11.0"
httpcore = "1.0.5"
httpx = "0.27.2"
huggingface-hub = "0.25.1"
hypothesis = "6.112.1"
idna = "3.10"
imageio = "2.35.1"
imageio-ffmpeg = "0.5.1"
inflection = "0.5.1"
inquirerpy = "0.3.4"
ipykernel = "6.29.5"
ipython = "8.27.0"
ipywidgets = "8.1.5"
isoduration = "20.11.0"
itsdangerous = "2.2.0"
jedi = "0.19.1"
Jinja2 = "3.1.4"
joblib = "1.4.2"
json5 = "0.9.25"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
jupyter = "1.1.1"
jupyter-console = "6.6.3"
jupyter-events = "0.10.0"
jupyter-lsp = "2.2.5"
jupyter_client = "8.6.3"
jupyter_core = "5.7.2"
jupyter_server = "2.14.2"
jupyter_server_terminals = "0.5.3"
jupyterlab = "4.2.5"
jupyterlab_pygments = "0.3.0"
jupyterlab_server = "2.27.3"
jupyterlab_widgets = "3.0.13"
keras = "3.5.0"
kiwisolver = "1.4.7"
lazy_loader = "0.4"
libclang = "18.1.1"
librosa = "0.10.2.post1"
llvmlite = "0.43.0"
loguru = "0.7.2"
Markdown = "3.7"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
matplotlib = "3.9.2"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
mistune = "3.0.2"
ml-dtypes = "0.4.1"
mongoengine = "0.29.1"
moviepy = "1.0.3"
mpmath = "1.3.0"
msgpack = "1.1.0"
multidict = "6.1.0"
multiprocess = "0.70.16"
namex = "0.0.8"
nbclient = "0.10.0"
nbconvert = "7.16.4"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
networkx = "3.3"
notebook = "7.2.2"
notebook_shim = "0.2.4"
numba = "0.60.0"
numpy = "1.26.4"
opencv-contrib-python = "4.10.0.84"
opencv-python = "4.10.0.84"
opt-einsum = "3.3.0"
optree = "0.12.1"
overrides = "7.7.0"
packaging = "24.1"
pandas = "2.2.3"
pandocfilters = "1.5.1"
parso = "0.8.4"
pfzy = "0.3.4"
pillow = "10.2.0"
platformdirs = "4.3.6"
pooch = "1.8.2"
proglog = "0.1.10"
prometheus_client = "0.21.0"
prompt_toolkit = "3.0.47"
protobuf = "4.25.5"
psutil = "6.0.0"
pure_eval = "0.2.3"
py4j = "0.10.9.7"
pyarrow = "17.0.0"
pycparser = "2.22"
pyctcdecode = "0.5.0"
Pygments = "2.18.0"
PySocks = "1.7.1"
pygtrie = "2.5.0"
PyJWT = "2.9.0"
pymongo = "4.9.1"
pyparsing = "3.1.4"
PySoundFile = "0.9.0.post1"
pyspark = "3.5.3"
python-dateutil = "2.9.0.post0"
python-json-logger = "2.0.7"
pytz = "2024.2"
pywin32 = { version = "306", markers = "sys_platform == 'win32'" }
pywinpty = { version = "2.0.13", markers = "sys_platform == 'win32'" }
win32-setctime = { version = "1.1.0", markers = "sys_platform == 'win32'" }
ptyprocess = "0.7.0"
pexpect = "4.9.0"
pytest = "8.4.1"
pytest-django = "4.11.1"
python-decouple = "3.8"
python-dotenv = "1.0.1"
PyYAML = "6.0.2"
pyzmq = "26.2.0"
redis = "5.0.8"
referencing = "0.35.1"
regex = "2024.9.11"
requests = "2.32.3"
resampy = "0.4.3"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rich = "13.8.1"
rpds-py = "0.20.0"
safetensors = "0.4.5"
scikit-learn = "1.5.2"
scipy = "1.14.1"
Send2Trash = "1.8.3"
simplejson = "3.19.3"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soundfile = "0.12.1"
soupsieve = "2.6"
soxr = "0.5.0.post1"
spotipy = "2.24.0"
sqlparse = "0.5.1"
stack-data = "0.6.3"
sympy = "1.13.3"
tensorboard = "2.17.1"
tensorboard-data-server = "0.7.2"
tensorflow = "2.17.0"
tensorflow-intel = { version = "2.17.0", markers = "sys_platform == 'win32'" }
tensorflow-io-gcs-filesystem = "*"
termcolor = "2.4.0"
terminado = "0.18.1"
tf_keras = "2.17.0"
threadpoolctl = "3.5.0"
tinycss2 = "1.3.0"
tokenizers = "0.19.1"
tomli = "2.0.1"
torch = { version = "2.2.2", markers = "sys_platform != 'win32'" }
torchaudio = { version = "2.2.2", markers = "sys_platform != 'win32'" }
torchvision = { version = "0.17.2", markers = "sys_platform != 'win32'" }
tornado = "6.4.1"
tqdm = "4.66.5"
traitlets = "5.14.3"
transformers = "4.44.2"
types-python-dateutil = "2.9.0.20240906"
typing_extensions = "4.12.2"
tzdata = "2024.2"
uri-template = "1.3.0"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
wcwidth = "0.2.13"
whitenoise = "6.7.0"
webcolors = "24.8.0"
webencodings = "0.5.1"
websocket-client = "1.8.0"
Werkzeug = "3.0.4"
widgetsnbextension = "4.0.13"
wrapt = "1.16.0"
xxhash = "3.5.0"
yarl = "1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
